moved ContactForm component and styles module to separate folder
